agent, common: move input-parsers module to indexer-common
And reorganize parsers in different modules,so that implementation
details stay separate from the module's interface (validators).
